Personal-Safety-Device-main/hardware/comms.py
import serial
import time
import requests
import os

class SIM7600:
    def __init__(self, port, baud=115200, timeout=1):
        try:
            self.ser = serial.Serial(port, baud, timeout=timeout)
            print("SIM7600 module initialized.")
        except Exception as e:
            self.ser = None
            print(f"Error: SIM7600 serial not available - {e}")

    # ... (all other functions like _send_command, send_sms, etc. are unchanged) ...
    def _send_command(self, command, expected_response, timeout):
        if not self.ser:
            return False, "Not connected"
        
        self.ser.write((command + '\r\n').encode())
        start_time = time.time()
        response = ''
        while time.time() - start_time < timeout:
            response_line = self.ser.readline().decode('utf-8', errors='ignore')
            if response_line:
                response += response_line
            if expected_response in response:
                return True, response
        return False, response

    def send_sms(self, number, text):
        if not self.ser: return False
        try:
            self._send_command('AT+CMGF=1', 'OK', 1)
            cmd = f'AT+CMGS="{number}"'
            success, _ = self._send_command(cmd, '>', 2)
            if success:
                self.ser.write(text.encode() + b"\x1A")
                sms_success, _ = self._send_command('', 'OK', 20)
                if sms_success:
                    print("SMS sent successfully.")
                    return True
            print("Failed to send SMS.")
            return False
        except Exception as e:
            print(f"Error during send_sms: {e}")
            return False

    def place_call(self, number):
        if not self.ser: return False
        success, _ = self._send_command(f'ATD{number};', 'OK', 10)
        return success

    def hang_up_call(self):
        if not self.ser: return False
        return self._send_command('AT+CHUP', 'OK', 5)[0]

    def get_gps_location(self):
        if not self.ser: return None
        
        self._send_command('AT+CGPS=1,1', 'OK', 1)
        
        print("Acquiring GPS fix...")
        maps_link = None
        for i in range(15):
            success, response = self._send_command('AT+CGPSINFO', '+CGPSINFO:', 2)
            if success and ',,,,,,' not in response:
                try:
                    parts = response.split(': ')[1].split(',')
                    lat_raw, lat_dir, lon_raw, lon_dir = parts[0], parts[1], parts[2], parts[3]
                    lat_deg, lat_min = divmod(float(lat_raw), 100)
                    lon_deg, lon_min = divmod(float(lon_raw), 100)
                    latitude = lat_deg + (lat_min / 60)
                    longitude = lon_deg + (lon_min / 60)
                    if lat_dir == 'S': latitude = -latitude
                    if lon_dir == 'W': longitude = -longitude
                    
                    maps_link = f"https://maps.google.com/?q={latitude},{longitude}"
                    
                    print(f"GPS Fix acquired: {maps_link}")
                    break
                except (ValueError, IndexError):
                    continue
            time.sleep(2)
        
        if not maps_link:
            print("Failed to get GPS fix.")

        self._send_command('AT+CGPS=0', 'OK', 1)
        
        return maps_link
    
    def upload_alert(self, server_url, alert_object, file_path):
        """
        Uploads a file and sends the full alert row data from the database object.
        Includes DEBUG LOGGING to see exactly what is being sent.
        """
        if not self.has_internet():
            print("No internet connection.")
            return False, None
        try:
            # Prepare the data payload from the alert object
            data = {
                'device_id': alert_object.device_id,
                'timestamp': alert_object.timestamp.isoformat(),
                'call_placed_status': alert_object.call_placed_status,
                'guardian_sms_status': alert_object.guardian_sms_status,
                'location_sms_status': alert_object.location_sms_status,
                'location': alert_object.gps_location,
                # --- NEW FIELD ---
                'battery_percentage': alert_object.battery_percentage 
            }

            # --- DEBUG: PRINT EXACTLY WHAT IS BEING SENT ---
            print("\n" + "="*40)
            print(" [DEBUG LOG] DATA BEING SENT TO SERVER ")
            print("="*40)
            print(f" Device ID:  {data['device_id']}")
            print(f" Timestamp:  {data['timestamp']}")
            print(f" Location:   {data['location']}")
            # --- NEW DEBUG LINE ---
            print(f" Battery %:  {data['battery_percentage']}")
            print(f" Call Stat:  {data['call_placed_status']}")
            print(f" SMS Stat:   {data['guardian_sms_status']}")
            print("="*40 + "\n")
            # ------------------------------------------------

            with open(file_path, 'rb') as f:
                files = {'file': (os.path.basename(file_path), f)}
                r = requests.post(server_url, files=files, data=data, timeout=30)
            
            if r.status_code == 201:
                print(f"Successfully uploaded {os.path.basename(file_path)}.")
                try:
                    uploaded_filename = r.json().get('filename')
                    if not uploaded_filename:
                        uploaded_filename = os.path.basename(file_path)
                    return True, uploaded_filename
                except requests.exceptions.JSONDecodeError:
                    return True, os.path.basename(file_path)
            else:
                print(f"Failed to upload file. Server responded with {r.status_code}.")
                return False, None
        except Exception as e:
            print(f"An error occurred during upload: {e}")
            return False, None
    
    @staticmethod
    def has_internet():
        try:
            requests.get('https://www.google.com', timeout=5)
            return True
        except Exception:
            return False
