Personal-Safety-Device-main/requirements.txt
SQLAlchemy>=1.4
requests>=2.25
pyserial>=3.5
RPi.GPIO
smbus2